pyproject.toml
[project]
name = "gestalt-game-engine"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beartype>=0.20.0",
    "pre-commit>=4.1.0",
    "pygame-ce>=2.5.3",
    "pyinstaller>=6.12.0",
]

[dependency-groups]
dev = []
